+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - You can now edit Math formulae by clicking on them.
 - The tag count is now shown next to the tags in the tag cloud.
 - During global search, the search results will include all files once at most, so files within virtual directories, for example, will be excluded to prevent duplicate files.
+- The preview images when pasting an image from clipboard load faster.
 
 ### Fixes